fix(discover-viability-gate): honor Groom-pass resolved-decision lines
The A4 autonomous_completion criterion discarded an issue exactly after
it was groomed per docs/idd-workflow.md's own instructions: recording
`Maintainer decision (<provenance>, <date>): <resolution text>` in the
body. EXTERNAL_COORDINATION_PATTERN's `maintainer decision` bigram had
no exemption for this shape, even though suitability-triage.mts's Check
7 already recognizes it as a resolved decision (#2661/#2711) through a
hardened, multi-round-reviewed regex and framing-verb/blockquote/
strikethrough exclusion scan. Reproduced with the exact Groom-pass line
plus a checklist acceptance criterion: A4 failed autonomous_completion
even though Check 7 would have honored the same line.

Rather than add a second, independently-drifting regex to
discover-viability-gate.mts, extract Check 7's resolved-decision
detector into a new shared module, src/scripts/resolved-decision.mts,
and point both files at it. suitability-triage.mts's own Check 7 now
calls the imported `hasResolvedDecision` instead of a local copy (no
behavior change -- the full test suite, including
tests/suitability-triage.test.mts, passes unchanged). Autonomous
completion now excludes an EXTERNAL_COORDINATION_PATTERN match that
falls inside a genuine resolved-decision span, using the same
per-occurrence exclusion structure the other autonomous_completion
exclusions already use.

Adds regression coverage for the exact reproduction (now passes), a
bare "needs a maintainer decision" mention with no resolved-decision
shape (still fails), and a quoted example of the decision line (keeps
its current, already-passing result).

Closes #2763
Refs #2661, #2738
